requirements.txt
python-dotenv>=1.0.0
pydantic>=2.7.0
requests==2.32.2
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
lumibot>=2.6.0
rich>=13.7.0
streamlit>=1.36.0
yfinance>=0.2.40

